+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes to this project.
 - `download_dataset` no longer requires a destination path, it defaults to the source file name
 - `download_dataset` now accepts a `round_num` argument, to download old files
 - added `upload_diagnostics` to upload to the new diagnostics tool
+- added `diagnostics` to fetch results of a diagnostics run
 
 ## [2.8.1] - 2021-09-08
 -  Add version arg to upload_predictions (#59)